Created a fresh Visual Studio Solution and Added some extra classes and features
Created Fresu Visual Studio solution since previous solution required Utilities which was in some other project which caused errors when running the Sockets project on it's own.

Extracted all the common imports and added some utility methods for Server and Client classes in SocketCommons.h file.

Previously utilities was added to the project but I forgot to include it. Including it in this project and tested. Everything works as intended.